pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "trade-agent"
dynamic = ["version"]
authors = [
  { name="Trading RL Platform Team", email="contact@trading-rl-platform.org" },
]
description = "A trading agent for reinforcement learning."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    # Core trading and ML dependencies
    "alpaca-py>=0.13.1",
    "numpy<2.0.0,>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "torch==2.8.0",
    "gymnasium==1.0.0",
    "ray[rllib,tune]==2.48.0",
    "yfinance>=0.2.0",
    "stable-baselines3[extra]==2.7.0",
    "lightgbm>=3.3.0",
    "html5lib>=1.1.0",
    "beautifulsoup4>=4.12.0",

    # Scientific computing and analysis
    "scipy>=1.10.0",
    "matplotlib>=3.6.0",
    "psutil>=5.9.0",
    "optuna>=2.10.0, <3.0.0",
    "xgboost>=1.6.0",
    "tsfresh>=0.18.0",
    "lxml>=0.9.0",
    "hexdump>=3.0.0",


    # Visualization
    "plotly>=5.0.0",
    "seaborn>=0.12.0",
    "tensorboard>=2.12.0",

    # Database dependencies
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",

    # Documentation dependencies
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.2.0",
    "myst_parser>=2.0.0",
    "sphinx_autodoc_typehints>=1.24.0",
    "sphinxcontrib-mermaid>=0.9.2",
    "sphinx-multiversion>=0.2.4",

    # Code quality and formatting
    "ruff>=0.1.0",
    "black>=23.3.0",
    "isort>=5.10.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
    "autoflake>=2.0.0",
    "pydocstyle>=6.3.0",
    "pyupgrade>=3.10.0",
    "pre-commit>=3.0.0",
    "bandit>=1.7.0",

    # Testing dependencies
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "coverage>=7.0.0",
    "hypothesis>=6.80.0",
    "locust>=2.15.0",
    "mutmut>=2.4.0",
    "interrogate>=1.5.0",
    "radon>=6.0.1",
    "pip-audit>=2.6.0",

    # Notebook support
    "nbqa>=1.7.0",

    # Type stubs
    "types-PyYAML>=6.0.0",
    "types-requests>=2.31.0",
    "types-setuptools>=68.0.0",

    # Core system
    "pylint>=2.16.0",
    # Configuration & hyperparameter optimization
    "hydra-core>=1.3.2",
    "hydra-optuna-sweeper>=1.2.0",
    # Data pipeline dependencies
    "polars>=0.20.0",
    "pydantic>=2.0.0",
    "tenacity>=8.2.0",
    # API service layer
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    # Experiment tracking
    "mlflow>=2.0.0",
]

# Expose extras for development (install with: pip install -e .[dev])
[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "mypy",
    "ruff",
    "black",
    "isort",
    "pre-commit",
]

[project.entry-points."trade_agent.data_collectors"]
memory = "trade_agent.plugins.builtins:simple_memory_collector"

[project.entry-points."trade_agent.rl_agents"]
hybrid = "trade_agent.plugins.builtins:make_hybrid_agent"

[project.entry-points."trade_agent.brokers"]
in_memory = "trade_agent.plugins.builtins:InMemoryBroker"


[tool.setuptools.packages.find]
where = ["src"]
include = ["trade_agent*"]
namespaces = false

[tool.black]
line-length = 88
target-version = ["py311"]
skip-string-normalization = true

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E","F","W",      # pyflakes/pycodestyle
    "I",                # isort
    "B",                # bugbear
    "UP",               # pyupgrade
    "N",                # pep8-naming
    "C90",              # mccabe complexity
    "DTZ",              # datetime
    "T20",              # print statements
    "ANN",              # type annotations
    "RET",              # return
    "SIM"               # simplifications
]
ignore = ["ANN101","ANN102","ANN401"]  # allow missing self/cls, Any in some contexts
exclude = ["docs/_build", "build", "dist"]

[tool.ruff.isort]
known-first-party = ["trade_agent"]
force-single-line = false
combine-as-imports = true
lines-after-imports = 2

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
disallow_any_unimported = false
ignore_missing_imports = true
show_error_codes = true
exclude = ["docs/_build", "build", "dist"]

#[tool.coverage.run] already defined above (deduplicated)
#[tool.coverage.report] already defined above (deduplicated)

#[tool.pytest.ini_options] duplicated section removed above

[project.urls]
"Homepage" = "https://github.com/trading-rl-platform/trading-rl-agent"
"Bug Tracker" = "https://github.com/trading-rl-platform/trading-rl-agent/issues"

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "rl: marks tests as reinforcement learning tests",
    "integration: marks tests as integration tests"
]
testpaths = ["tests"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov-report=term-missing:skip-covered",
    "--cov-fail-under=80"
]
